+++ libcore/tests/crypto/rpo.rs
@@ -19,7 +19,7 @@ fn test_invalid_end_addr() {
 
     expect_exec_error_matches!(
         test,
-        ExecutionError::FailedAssertion{ clk, err_code, err_msg , label: _, source_file: _ }
+        ExecutionError::FailedAssertion{ clk, err_code, err_msg , .. }
         if clk == RowIndex::from(24) && err_code == ZERO && err_msg.is_none()
     );
 }