+++ processor/src/fast/mod.rs
@@ -544,16 +544,23 @@ impl FastProcessor {
         match decorator {
             Decorator::Debug(options) => {
                 if self.in_debug_mode {
+                    let clk = self.clk;
                     let process = &mut self.state();
-                    host.on_debug(process, options)?;
+                    host.on_debug(process, options)
+                        .map_err(|err| ExecutionError::DebugHandlerError { clk, err })?;
                 }
             },
             Decorator::AsmOp(_assembly_op) => {
                 // do nothing
             },
             Decorator::Trace(id) => {
+                let clk = self.clk;
                 let process = &mut self.state();
-                host.on_trace(process, *id)?;
+                host.on_trace(process, *id).map_err(|err| ExecutionError::TraceHandlerError {
+                    clk,
+                    trace_id: *id,
+                    err,
+                })?;
             },
         };
         Ok(())