+++ processor/src/lib.rs
@@ -66,7 +66,10 @@ pub use host::{
     advice::{AdviceError, AdviceInputs, AdviceProvider},
     debug::DefaultDebugHandler,
     default::{DefaultHost, HostLibrary},
-    handlers::{DebugHandler, EventError, EventHandler, EventHandlerRegistry, NoopEventHandler},
+    handlers::{
+        AssertError, DebugError, DebugHandler, EventError, EventHandler, EventHandlerRegistry,
+        NoopEventHandler, TraceError,
+    },
 };
 
 mod chiplets;
@@ -665,7 +668,9 @@ impl Process {
             Decorator::Debug(options) => {
                 if self.decoder.in_debug_mode() {
                     let process = &mut self.state();
-                    host.on_debug(process, options)?;
+                    let clk = process.clk();
+                    host.on_debug(process, options)
+                        .map_err(|err| ExecutionError::DebugHandlerError { clk, err })?;
                 }
             },
             Decorator::AsmOp(assembly_op) => {
@@ -676,7 +681,10 @@ impl Process {
             Decorator::Trace(id) => {
                 if self.enable_tracing {
                     let process = &mut self.state();
-                    host.on_trace(process, *id)?;
+                    let clk = process.clk();
+                    host.on_trace(process, *id).map_err(|err| {
+                        ExecutionError::TraceHandlerError { clk, trace_id: *id, err }
+                    })?;
                 }
             },
         };