+++ processor/src/processor/operations/sys_ops.rs
@@ -21,9 +21,10 @@ pub(super) fn op_assert<P: Processor>(
 ) -> Result<(), ExecutionError> {
     if processor.stack().get(0) != ONE {
         let process = &mut processor.state();
-        host.on_assert_failed(process, err_code);
+        let clk = process.clk();
+        let err = host.on_assert_failed(process, err_code);
         let err_msg = program.resolve_error_message(err_code);
-        return Err(ExecutionError::failed_assertion(process.clk(), err_code, err_msg, err_ctx));
+        return Err(ExecutionError::failed_assertion(clk, err_code, err_msg, err, err_ctx));
     }
     processor.stack().decrement_size(tracer);
     Ok(())